IT Ticket and Work Order Management System.
Work Horse is an line of business web application that is used to
create, edit, and manage tickets and work orders for your companys IT
staff. It is accessible on Internet/Intranet and for technician
workstations, there is a companion desktop application, also called Work
Horse Lite, that pulls tickets from the database and shows which have
been worked and which need to be worked. This is ideal for help desk
support being able to check the status and make mild adjustments but
have the ability to escalate and communicate with senior level
IT/escalation. The companion application will also offer features
similar to the web application but will not offer full line of features.
The desktop application can be deployed to Windows 7/Windows 8.1/Windows
Server 2008 R2, Windows Server 2012 R2 clients. The web application will
be based on ASP.NET MVC 5 and Entity Framework 6. Data will be stored in
a SQL database and will implement a RESTful service to facilitate
communication for the requesting clients. The desktop application is
engineered using Windows Presentation Foundation 4.5/C# and will of
course be using the web service and SQL database/EF 6. Documentation
will be uploaded shortly. You can check out my blog for more updates.
http://fromtechtodev.blogspot.com ----
http://derekpauley.azurewebsites.net